Separate LedArray code into another file
